third_party/blink/web_tests/platform/mac-mac-arm11.0/external/wpt/css/css-fonts/generic-family-keywords-001-expected.txt

This is a testharness.js-based test.
PASS @font-face matching for quoted and unquoted serif
PASS @font-face matching for quoted and unquoted sans-serif
PASS @font-face matching for quoted and unquoted cursive
PASS @font-face matching for quoted and unquoted fantasy
PASS @font-face matching for quoted and unquoted monospace
FAIL @font-face matching for quoted and unquoted system assert_equals: unquoted system does not match @font-face rule expected 25 but got 50
FAIL @font-face matching for quoted and unquoted emoji assert_equals: unquoted emoji does not match @font-face rule expected 25 but got 50
FAIL @font-face matching for quoted and unquoted math assert_equals: unquoted math does not match @font-face rule expected 25 but got 50
FAIL @font-face matching for quoted and unquoted fangsong assert_equals: unquoted fangsong does not match @font-face rule expected 25 but got 50
FAIL @font-face matching for quoted and unquoted ui-serif assert_equals: unquoted ui-serif does not match @font-face rule expected 25 but got 50
FAIL @font-face matching for quoted and unquoted ui-sans-serif assert_equals: unquoted ui-sans-serif does not match @font-face rule expected 25 but got 50
FAIL @font-face matching for quoted and unquoted ui-monospace assert_equals: unquoted ui-monospace does not match @font-face rule expected 25 but got 50
FAIL @font-face matching for quoted and unquoted ui-rounded assert_equals: unquoted ui-rounded does not match @font-face rule expected 25 but got 50
Harness: the test ran to completion.
